Reorganize top level and utils files
