+++ src/logic/__tests__/scope.ts
