+++ src/plugins/documentation/convert.ts
@@ -3,7 +3,7 @@ import * as serialization from '@lona/serialization'
 import { Helpers } from '../../helpers'
 import { convertDeclaration } from '../tokens/convert'
 import { Token } from '../tokens/tokensAst'
-import { nonNullable, assertNever } from '../../utils'
+import { assertNever, nonNullable } from '../../utils/typeHelpers'
 
 let tokenNameElement = (kind: string, content: string) =>
   `<span class="lona-token-name lona-token-name-${kind}">${content}</span>`
