--- src/utils/index.ts
@@ -1,9 +1,0 @@
-export * from './findPlugin'
-export * from './isWorkspacePath'
-export * from './findWorkspace'
-export * from './nonNullable'
-export * from './typeHelpers'
-export * from './shallowMap'
-export * from './uuid'
-import * as config from './config'
-export { config }
