--- src/utils/nonNullable.ts
@@ -1,3 +1,0 @@
-export function nonNullable<T>(value: T): value is NonNullable<T> {
-  return value !== null && value !== undefined
-}
